+++ src/main/java/ch/furchert/homelab/auth/dto/LoginResponse.java
@@ -1,6 +1,0 @@
-package ch.furchert.homelab.auth.dto;
-
-public record LoginResponse(
-        String accessToken,
-        String refreshToken
-) {}